+++ packages/sdk/src/sync/subgraph/core/adapter.ts
@@ -92,10 +92,19 @@ export class SubgraphSDKSyncAdapter implements SDKSyncAdapter {
           yield stateDiff;
         } else {
           // otherwise, we've caught up and there's nothing more to fetch.
-          // set `from` to the entity index corresponding to the latest indexed block and continue to avoid emitting empty diff
-          from = TotalEntityIndexTrait.fromComponents({
-            blockNumber: BigInt(currentBlock),
-          });
+          // set `from` to the entity index corresponding to the latest indexed block
+          // if it's greater than the current `from`.
+
+          // this is to prevent an busy loops in the case where the subgraph has indexed a block corresponding
+          // to a totalEntityIndex > `endTotalEntityIndex` but we haven't found any insertions in that block
+          const currentBlockTotalEntityIndex =
+            TotalEntityIndexTrait.fromComponents({
+              blockNumber: BigInt(currentBlock),
+            });
+
+          if (currentBlockTotalEntityIndex > from) {
+            from = currentBlockTotalEntityIndex;
+          }
         }
       }
     };