+++ packages/subtree-updater/src/subtreeUpdater.ts
@@ -156,25 +156,22 @@ export class SubtreeUpdater {
     const queuerProm = queuer();
 
     const teardown = async () => {
-      const results = await Promise.allSettled([
+      await Promise.allSettled([
         proofJobs.close(),
         prover.close(),
         submitter.close(),
       ]);
 
-      const teardownResults = {
-        closeProofJobs: results[0],
-        closeProver: results[1],
-        closeSubmitter: results[2],
-      };
-      this.logger.debug(`teardown completed with results`, { teardownResults });
+      await Promise.allSettled([submitterProm, proverProm, queuerProm]);
+
+      this.logger.debug("teardown completed");
     };
 
     const promise = (async () => {
       try {
         await Promise.all([submitterProm, proverProm, queuerProm]);
       } catch (err) {
-        this.logger.error(`error in subtree updater: ${err}`, err);
+        this.logger.error(`error in subtree updater`, err);
         await teardown();
         throw err;
       }
@@ -274,7 +271,6 @@ export class SubtreeUpdater {
         logger.debug(`got insertion at merkleIndex ${merkleIndex}`, {
           insertion,
         });
-        merkleIndex += 1;
         if (this.fillBatchLatency === undefined) {
           return;
         }
@@ -288,7 +284,7 @@ export class SubtreeUpdater {
         clearTimeout(this.fillBatchTimeout);
 
         // if the insertion we got is not at a batch boundry, re-set the timeout because we haven't organically filled the batch yet
-        if (merkleIndex % BATCH_SIZE !== 0) {
+        if ((merkleIndex + 1) % BATCH_SIZE !== 0) {
           this.fillBatchTimeout = setTimeout(
             () =>
               this.fillBatchWithZeros(